Task#43 Scheduler for chat polls Added after each method to truncate the db Separated IT extensions for packages Refactored
