+++ core/src/integrationTest/java/by/mrrockka/repo/entry/EntriesRepositoryTest.java
@@ -1,7 +1,7 @@
 package by.mrrockka.repo.entry;
 
 import by.mrrockka.IntegrationTestConfiguration;
-import by.mrrockka.config.PostgreSQLExtension;
+import by.mrrockka.config.CorePSQLExtension;
 import by.mrrockka.repo.entries.EntriesEntity;
 import by.mrrockka.repo.entries.EntriesRepository;
 import by.mrrockka.repo.person.PersonRepository;
@@ -17,7 +17,7 @@ import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ExtendWith(PostgreSQLExtension.class)
+@ExtendWith(CorePSQLExtension.class)
 @SpringBootTest(classes = IntegrationTestConfiguration.class)
 class EntriesRepositoryTest {
   private static final UUID GAME_ID = UUID.fromString("fa3d03c4-f411-4852-810f-c0cc2f5b8c84");
