+++ core/src/integrationTest/java/by/mrrockka/repo/prizepool/PrizePoolRepositoryTest.java
@@ -1,7 +1,7 @@
 package by.mrrockka.repo.prizepool;
 
 import by.mrrockka.IntegrationTestConfiguration;
-import by.mrrockka.config.PostgreSQLExtension;
+import by.mrrockka.config.CorePSQLExtension;
 import by.mrrockka.creator.PrizePoolCreator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -12,7 +12,7 @@ import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ExtendWith(PostgreSQLExtension.class)
+@ExtendWith(CorePSQLExtension.class)
 
 @SpringBootTest(classes = IntegrationTestConfiguration.class)
 class PrizePoolRepositoryTest {
