+++ telegram/src/testFixtures/java/by/mrrockka/config/TelegramPSQLExtension.java
@@ -1,0 +1,21 @@
+package by.mrrockka.config;
+
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.sql.PreparedStatement;
+
+@Slf4j
+public class TelegramPSQLExtension extends CorePSQLExtension {
+
+  @Override
+  public void afterEach(final ExtensionContext context) {
+//    todo: generate data instead of migration scripts usage
+    final var jdbcTemplate = SpringExtension.getApplicationContext(context).getBean(NamedParameterJdbcTemplate.class);
+//    jdbcTemplate.execute("truncate chat_persons", PreparedStatement::execute);
+//    jdbcTemplate.execute("truncate chat_games", PreparedStatement::execute);
+    jdbcTemplate.execute("truncate poll_task", PreparedStatement::execute);
+  }
+}
